% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateOTUs.R
\name{aggregateOTUs}
\alias{aggregateOTUs}
\title{Aggregate OTUs on different taxonomic levels}
\usage{
aggregateOTUs(otu, tax, level = 6)
}
\arguments{
\item{otu}{a table of OTUs (columns) with samples in rows}

\item{tax}{a table with classification (levels {kingdom, phylum, etc} in columns and OTUs in rows)}

\item{level}{Which level to aggregate on (corresponding to column in `tax`, defaults to 6 which normally corresponds to genus)}
}
\value{
A list with:

`tax` Taxonomy up to the level of aggregation

`name` Full name of aggregated variables

`count` A table with aggregated counts
}
\description{
Aggregate OTUs on different taxonomic levels
}
\examples{
# genus <- aggregateOTUs(inData,taxonomyDF,6)
}
