% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{binlr_nll}
\alias{binlr_nll}
\title{Negative log likelihood for binary logistic regression}
\usage{
binlr_nll(par, D, y)
}
\arguments{
\item{par, }{a vector}

\item{D, }{data in the form of a data frame, matrix or vector}

\item{y, }{a vector where entries are 0 or 1}
}
\value{
a numeric
}
\description{
Given vector of parameters input data, D, and classes for the input data, y, computes and returns the negative log likelihood,
}
\examples{
x <- c(1:10)
y <- c(rep(0,5), rep(1,5))
par <- c(1,2)
binlr_nll(par, x, y)
# To compute the MLE:
optim(par = c(0,0), fn = neg_log_likelihood, D=x, y=y)
}
