% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforS3methods.R
\name{checkforS3methods}
\alias{checkforS3methods}
\title{checkforS3methods}
\usage{
checkforS3methods(func)
}
\arguments{
\item{func}{A data frame with variables with class "numeric", "factor", "logical","character","ordered factor"}
}
\value{
A list of S3 methods cthat can match this function
}
\description{
checkforS3methods
}
\examples{
checkforS3methods("predict")
}
