% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkforsanity.R
\name{checkforsanity}
\alias{checkforsanity}
\title{Check for sanity of a dataset}
\usage{
checkforsanity(
  data,
  check = c("NA", "NaN", "negative", "0", "infinite"),
  output = c("number", "percentage")
)
}
\arguments{
\item{data}{The dataset used for checking. It could be a dataframe or matrix}

\item{check}{"NA","NaN","negative",0,infinite. The default value is NA}

\item{output}{"number","percentage". The default value}
}
\value{
A table: The first row is how many values is "NA","NaN","negative",0,infinite. The second row is how many columns has this value
}
\description{
Check for sanity of a dataset
}
