% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{df_to_numeric}
\alias{df_to_numeric}
\title{Turns data.frame into a numeric}
\usage{
df_to_numeric(df)
}
\arguments{
\item{df, }{data.frame}
}
\value{
df, numeric
}
\description{
Given a data.frame df, will convert to a numeric and return it.
}
\examples{
df_to_numeric(cars)
}
