% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-seasonal-temperatures.R
\name{dSin}
\alias{dSin}
\title{Probability density of scaled sine wave}
\usage{
dSin(x, mean.t, amp.t)
}
\arguments{
\item{x}{numeric; vector of values for which to obtain the probability
densities.}

\item{mean.t}{numeric; mean of the sine wave.}

\item{amp.t}{numeric; peak-to-peak amplitude of the sine wave.}
}
\value{
Numeric vector of probability density values for the values given in
  \code{x}. Values in \code{x} outside the sine wave's envelope obtain zero
  probability.
}
\description{
Calculate the probability density for a scaled sine wave.
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
