% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{E_l2}
\alias{E_l2}
\title{L2 Norm Error}
\usage{
E_l2(targ, pred)
}
\arguments{
\item{targ, }{a vector}

\item{pred, }{a vector}
}
\value{
a numeric
}
\description{
Given targets and predictions computes the L2 norm of their difference.
}
\examples{
err <- E_l2( c(5,4,3,2,1), c(6,3,4,1,2) )
}
