% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-functions.R
\name{ExpectedRange}
\alias{ExpectedRange}
\title{Expected range}
\usage{
ExpectedRange(sd, n)
}
\arguments{
\item{sd}{standard deviation}

\item{n}{number of observations}
}
\value{
numeric
}
\description{
Calculates the expected range of n observations drawn from a distibution with standard deviation sd
}
\examples{

ExpectedRange(5, 15)

\dontrun{
library(tidyverse)
f <- function(x, n) n * x * pnorm(x)^(n - 1) * dnorm(x)
ests <- expand.grid(mean = 0, sd = 2, n = 15, rep = 1:1000, i = 1:15) \%>\%
  mutate(rdv = rnorm(n(), mean, sd)) \%>\%
  group_by(mean, sd, n, rep) \%>\%
  summarise(sd_hat = sd(rdv),
            obs.range = diff(range(rdv))) \%>\%
  group_by(mean, sd, n, rep) \%>\%
  mutate(sd.range = 2*sd_hat*integrate(f,-Inf,Inf, n = n)$value,
         exp.range = ExpectedRange(sd, n))
ests \%>\%
  gather(estimate, value, -mean, -sd, -n, -rep, -sd_hat, -exp.range) \%>\%
  ggplot(aes(x = estimate, y = value)) +
  geom_boxplot() +
  geom_hline(aes(yintercept = exp.range)) +
  facet_wrap(~n)
  }
}
