% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools_code_BIVAR.R
\name{export}
\alias{export}
\alias{export.bivarTable}
\alias{export.default}
\title{Export a matrix to tex, csv or xlsx}
\usage{
export(X, ...)

\method{export}{default}(X, cols = colnames(X), rows = rownames(X),
  type = "xlsx", file = tempfile(pattern = "exportDefault", fileext =
  paste0(".", type)), caption = NULL, label = NULL, append = FALSE, ...)

\method{export}{bivarTable}(X, type = "xlsx", file = tempfile(pattern =
  "exportBivarTable", fileext = paste0(".", type)), caption = NULL,
  label = NULL, ...)
}
\arguments{
\item{X}{Object to be exported (matrix or bivarTable object).}

\item{...}{Further arguments passed to print.xtable}

\item{cols}{Names for the columns}

\item{rows}{Names for the rows}

\item{type}{Type of file to export to (extension without the dot).}

\item{file}{Name of the file. By default a temp file is created.}

\item{caption}{Caption passed to xtable in case of exporting to tex.}

\item{label}{Label when exporting to tex.}

\item{append}{Append or not in case of csv.}
}
\value{
Returns the argument X but invisible.
}
\description{
Better visualisation of a matrix, or saving it in a file.
}
\details{
When exporting to xlsx, the produced file is opened. Make sure openxlsx is working ok on your computer (Rtools and PATH issues).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for export.

\item \code{bivarTable}: For exporting a bivarTable object, with fancier style if type is xlsx.
}}
\seealso{
\code{\link{bivarTable}}
}

