% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_samesequence.R
\name{factor_samesequence}
\alias{factor_samesequence}
\title{This function is to check for overlapping of values or dataframe columns in different objects.
You could see this as a venn diagram but in nicer text version. It does not have the limitation of venn diagram that only at most 5-6 object can be shown together}
\usage{
factor_samesequence(X, level, sequence)
}
\arguments{
\item{X}{a factor variable}

\item{level}{levels if want to manually set}

\item{sequence}{default as F, if T normal factor}
}
\value{
X_factor
}
\description{
This function is to check for overlapping of values or dataframe columns in different objects.
You could see this as a venn diagram but in nicer text version. It does not have the limitation of venn diagram that only at most 5-6 object can be shown together
}
