% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{feat_trans}
\alias{feat_trans}
\title{Performs a polynomial feature transform}
\usage{
feat_trans(x, b)
}
\arguments{
\item{x, }{a numeric or vector}

\item{b, }{a numeric}
}
\value{
x_ft, an array (either 1 or 2 dimensional)
}
\description{
Given data, x, and degree of polynomial transform to perform, b, computes and returns the polynomial feature transform of degree b on x, x_ft.
}
\examples{
feat_trans(5,3)
feat_trans(c(1,2,3,4), 4)
}
