% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterClass.R
\name{filterClass}
\alias{filterClass}
\title{Filter out features based on presence in classes}
\usage{
filterClass(table, classes, ratio = 0.5)
}
\arguments{
\item{table}{Feature matrix (samples in rows, features in columns)}

\item{classes}{Vector of sample classes}

\item{ratio}{Ratio for decision keep/discard}
}
\value{
Feature table filtered by class presence
}
\description{
To be kept, a feature need to be present in ≥`ratio` in ≥1 class
}
