% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency table for a single variable}
\usage{
freq(var, row, colnames)
}
\arguments{
\item{var}{A categorical variable to be analyzed}

\item{row}{Row name for the output}

\item{colnames}{Level names for the categorical variable}
}
\description{
This is a frequency table function which prints the frequency
 and related row percentage for a given categorical variable.The function
  creates the count and related row percentages for a given
categorical variables and allows naming the header.
Provide the distinct level names as a list.
}
\examples{
#Example
set.seed(1234)
letter = list(sample(letters[1:5], 20, replace=TRUE))
freq( letter, row = c("Letter"), colnames = c("A", "B", "C", "D", "E"))
}
