% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcdnum}
\alias{gcdnum}
\title{Codes for Austrian municipalities}
\format{A data frame with 2379 rows and 5 variables:
\itemize{
  \item GKZ11: municipality codes until 2011
  \item GKZ12: municipality codes 2012 (new district MT)
  \item GKZ13: municipality codes 2013 to 2014 (new districts
BM, HF, SO)
  \item GKZ15: municipality codes 2015 (changes in Styrian
municipalities)
  \item GKZ16: municipality codes 2016 (changes in
Upper Austria)
  \item GKZ17: municipality codes 2017 (changes in
Lower Austria)
}}
\usage{
data("gcdnum")
}
\description{
A dataset containing the codes for municipalities in Austria
from 2011 to 2016 (estimated) -- Gemeindecodes (GCD)
}
\keyword{datasets}

