% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdrecode.R
\name{gcdrecode}
\alias{gcdrc}
\alias{gcdrecode}
\title{Recode Austrian "Gemeindecodes"}
\usage{
gcdrecode(year, gcd, targetyear = 2017, invalid = "NA")

gcdrc(year, gcd, targetyear = 2017, ...)
}
\arguments{
\item{year}{a character or an integer containing the year}

\item{gcd}{a vector containing Austrian "Gemeindekennziffern"
corresponding to the \code{year}}

\item{targetyear}{a character or an integer containing the
year for which the \code{gcd} should be
returned}

\item{invalid}{a character or an integer specifying what
should be returned for invalid
gcd-year-combinations (see details)}

\item{\dots}{further arguments in \code{gcdrc} passed on
\code{gcdrecode}}
}
\value{
a vector containing Austrian "Gemeindecodes"
        corresponding to the \code{targetyear}
}
\description{
Recode Austrian "Gemeindecodes" due to changes
over the years (e.g. "Gemeindestrukturreform 2015")
}
\details{
Different states for \code{gcd} are available for 2011, 2012,
2013, 2015, 2016 and 2017 (first of January). All of them,
except the first one, are therfore meaningful values for
\code{targetyear}, which has to be higher than \code{year}.

\code{gcdrc} is the vectorized version to be used for
different values of \code{year}. If \code{year} is constant
in your data, \code{gcdrecode} should be used to avoid
needless overhead.

There are three possibilities to handle wrong
gcd-year-combinations, which can be caused by wrong gcd e.g.
if the last two digits are zero (60100):
\enumerate{
  \item \code{invalid = "NA"} (default): result is set to
        \code{\link{NA}}
  \item \code{invalid = "orig"}: original values of
        \code{gcd} are returned instead of recoding them
  \item \code{invalid = <something else>}: something else
        (character or numeric) is returned
}

For some \code{gcd} there is no unique match possible due to
splitting of municipalities in 2015. For these cases the
\code{gcd} of the part with the higher population (2015) is
used and the other is skipped (\code{\link{warning}}). The
numbers in brackets are the \code{gcd} belonging to the names
of municipalities in 2013 (before splitting) respectively
2015 (after splitting).

\itemize{
  \item Kohlberg (62336)
    \itemize{
      \item used Gnas (62380)
      \item skipped Paldau (62384)
    }
  \item Oberstorcha (62349)
    \itemize{
      \item used Paldau (62384)
      \item skipped Kirchberg an der Raab (62382)
    }
  \item Limbach bei Neudau (62277)
    \itemize{
      \item used Neudau (62274)
      \item skipped Bad Waltersdorf (62264)
    }
  \item Schlag bei Thalberg (62248)
    \itemize{
      \item used Rohrbach an der Lafnitz (62277)
      \item skipped Dechantskirchen (62265)
    }
  \item Stocking (61040)
    \itemize{
      \item used Wildon (61059)
      \item skipped Sankt Georgen an der Stiefing (61055)
    }
}
}
\examples{

### 62380 is existing since 2015
gcdrecode(2011, 62380, 2015) # warning
gcdrecode(2011, 62380, 2015, invalid = "orig") # warning
gcdrecode(2011, 62380, 2015, invalid = "wrong") # warning
gcdrecode(2015, 62380, 2017) # didn't change
gcdrc(2011:2012, c(60201, 60201), 2016)
### some gcd of the sequence are not existing --> warning
gcdrc(2016, 32401:32424, 2017) # warning; WU ex 2017
}
\author{
Gerhard Nachtmann \email{kpm.nachtmann@gmail.com}
}
\keyword{Gemeindecodes}
\keyword{gcdrc}
\keyword{gcdrecode}
\keyword{recode}

