% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAUROC.R
\name{getAUROC}
\alias{getAUROC}
\title{get miss classification from classification analysis}
\usage{
getAUROC(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual classifications of samples}

\item{predicted}{Vector of predicted classifications of samples}
}
\value{
Balanced Error Rate (BER)
}
\description{
get miss classification from classification analysis
}
