% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEffectiveDOF.R
\name{getEffectiveDOF}
\alias{getEffectiveDOF}
\title{Effective degrees of freedom}
\source{
Xu, X.: Methods in Hypothesis Testing, Markov Chain Monte Carlo and
  Neuroimaging Data Analysis, PhD thesis, Harvard University, Cambridge,
  Massachusetts, 2013, \url{https://dash.harvard.edu/handle/1/11108711}.
}
\usage{
getEffectiveDOF(n, a1)
}
\arguments{
\item{n}{number of observations.}

\item{a1}{lag 1 autocorrelation parameter.}
}
\value{
the effective degrees of freedom.
}
\description{
Calculate the effective degrees of freedom of an autocorrelated dataset
consisting of \code{n} observations, i.e. the nuber of independent
obervations given an autocorrelation parameter for the data estimated at lag
1.
}
\examples{
getEffectiveDOF(10, 0.6)
}
\author{
Thomas Laepple, Thomas Münch
}
