% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinclusionworklist.R
\name{getinclusionworklist}
\alias{getinclusionworklist}
\title{Please consult King Yan if smth is not clear}
\usage{
getinclusionworklist(
  Dir = getwd(),
  featurenames,
  oldfeaturename_format,
  newfeaturename_format,
  featurename_removerest = F,
  Dataframe_Obj,
  oldsamplename_format,
  newsamplename_format,
  samplename_removerest = F,
  splitmark = "@",
  rtOrNot = F,
  percentage = 0.1,
  chromPol,
  splitfeatures_withcloserententiontime,
  mergesamples_withfarthermz = T,
  instrumentName = c("Zoidberg", "Fry"),
  rententiontime_threshold = 20,
  samplesplitlistmark = "£",
  minute_to_second = F,
  batch = 1,
  Week = 1,
  Date,
  phase = "RP"
)
}
\arguments{
\item{Dir}{The directory where you want to save every thing of your output (It is different from the saveDir)}

\item{featurenames}{A vector of feature names. It must look like a format that is easy to change. for example RP_neg_674_3536a5_84 is okay}

\item{oldfeaturename_format}{a vector,what part in the feature name you want to change. Do not use $}

\item{newfeaturename_format}{a vector, what you want to change to for specific part of feature name, note that your new name should be Mass, splitmark, retention time. Do not use $}

\item{featurename_removerest}{logical, remove the rest of the name or not (the part in the name that is not changed)}

\item{Dataframe_Obj}{A DATAFRAME with rows as as sample and columns as covariates/features (even one column,it should be in dataframe format), which must contain the feature names you want to find,otherwise you get an error 
The rownames are names in a format that is easy to change, colnames are feature names}

\item{oldsamplename_format}{a vector,what part in the sample name you want to change. Do not use $}

\item{newsamplename_format}{a vector, what you want to change to for specific parts of sample name. Do not use $}

\item{samplename_removerest}{logical, remove the rest of the name or not (the part in the name that is not changed)}

\item{splitmark}{The mark you used to separate mass and retention time. For example "@"}

\item{rtOrNot}{Whether to use RT window when doing MS2 or not, default = False because usually we won't know where something is and there used to be trouble with the window historically}

\item{percentage}{What is the percentage threshold that you want to select relatively highest features}

\item{chromPol}{Either RN or RP}

\item{splitfeatures_withcloserententiontime}{logical, if true, use split sample feature function}

\item{mergesamples_withfarthermz}{To be developed, default 10}

\item{instrumentName}{Either Zoidberg or Fry
getinclusionworklist}

\item{rententiontime_threshold}{What difference value is considered as close. Usable when splitfeatures_withcloserententiontime is T}

\item{samplesplitlistmark}{A split mark for split samples. Usable when splitfeatures_withcloserententiontime is T}

\item{minute_to_second, }{logical, change retention time from minute to seconds}

\item{batch}{Batch on Worklist}

\item{Week}{Week on Worklist}

\item{Date}{Date on Worklist}

\item{phase}{Phase on Worklist}
}
\value{
sampleSet
}
\description{
Please consult King Yan if smth is not clear
}
