% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonomy.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Extract phylogenetic information at different taxonomic levels from character strings}
\usage{
getTaxonomy(taxon, split1 = ";", split2 = "__", depth = 6)
}
\arguments{
\item{taxon}{Character with OTU taxonomic information (e.g. "D_0__Bacteria;D_1__Firmicutes;D_2__Clostridia;D_3__Clostridiales;D_4__Lachnospiraceae;D_5__Blautia" ...)}

\item{split1}{Splitting character between taxonomic levels (e.g. ";" which is default)}

\item{split2}{Splitting character within taxonomic levels (e.g. "__" which is default)}

\item{depth}{At what taxonomic level to truncate (defaults to 6 which normally corresponds to genus)}
}
\value{
A data frame with taxonomic level information in columns
}
\description{
Often, OTUs are delivered with taxonomic information in string format:  
"D_0__Bacteria;D_1__Firmicutes;D_2__Clostridia;D_3__Clostridiales;D_4__Lachnospiraceae;D_5__Blautia" ...  
Rather than in speadsheets (data frames).  
This function provides a simple algorithm to extract phylogenetic information at different taxonomic levels 
and report them in a data frame format.
}
\examples{
# load(file='MBBL.rda')
# taxDF <- getTaxonomy(tax$Taxon)
}
