% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools_code_OTHER.R
\name{GoF}
\alias{GoF}
\title{Goodness of Fit p-value}
\usage{
GoF(model)
}
\arguments{
\item{model}{Model passed to \code{residuals(., "pearson")}.}
}
\value{
Resulting p-value from the test.
}
\description{
Assessing the fit of a model with Pearson's GoF statistic
}
\examples{
m <- glm(I(Sepal.Width > 3) ~ Species + Sepal.Length, iris, family = binomial)
GoF(m)

}
\seealso{
\code{\link{getPval}}, \code{\link{nagelkerke}}, \code{\link{getORCI}}, \code{\link{getBetaSd}},
\code{\link{verticalgetORCI}}, \code{\link{verticalgetBetaSd}}, \code{\link{verticalgetPval}}.
}

