% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyVec.R
\name{greyVec}
\alias{greyVec}
\title{Return a greyscale vector corresponding to values in vector (useful for plotting colors according to values)}
\usage{
greyVec(xCol, maxGrey = 85, minGrey = 10)
}
\arguments{
\item{xCol}{A vector with numerical values}

\item{maxGrey}{Max value for greyscale (defaults to 85)}

\item{minGrey}{Min value for greyscale (defaults to 10)}
}
\value{
A character vector with greyscale names
}
\description{
Return a greyscale vector corresponding to values in vector (useful for plotting colors according to values)
}
