% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_non_ordinal.R
\name{impute_non_ordinal}
\alias{impute_non_ordinal}
\title{imputation for non-ordinal factor variables}
\usage{
impute_non_ordinal(
  non_ordinal,
  guess = NULL,
  forceZero = FALSE,
  method = c("PLS", "RF"),
  rfMeth = c("rf", "Rborist", "ranger"),
  nComp = 2,
  nCore,
  tol1 = 0.05,
  n1 = 15,
  tol2 = 0.025,
  n2 = 60
)
}
\arguments{
\item{guess}{Initial guess (see mvImp; defaults to mvImp default)}

\item{forceZero}{Boolean for whether to force a lower imputation limit to zero (see mvImp; defaults to mvImp default)}

\item{method}{'PLS' or 'RF}

\item{rfMeth}{Which RF implementation to choose ('rf' (randomForest; default), 'ranger' or 'Rborist')}

\item{nComp}{Number of PLS components (defaults to 2)}

\item{nCore}{Number of slave processes (defaults to detectCores()-1)}

\item{tol1}{Tolerance in 1st iteration (defaults to 0.05)}

\item{n1}{MaxIter for 1st iteration (defaults to 15)}

\item{tol2}{Tolerance in 2nd iteration (defaults to 0.05)}

\item{n2}{MaxIter for 2nd iteration (defaults to 15)}

\item{non-ordinal}{A dataframe with non-ordinal variables}
}
\value{
A dataframe with imputed values
}
\description{
Preparation: To use this package, 1 steps need to be done beforehand:
}
\details{
Step 1: Made all variables that will be used a dataframe. For example, a data frame with name X

!!! Note: the class of variables in these data frames should be only "factor". (Use the Transform_fac_num in the same package.)

Step 2:  Make sure that the sequence of the levels for each factor variable become the "default sequence". (Use the Transform_fac_num in the same package.)
Example: a<-factor(c(1,2,3)). When levels(a), the out put is 1,2,3. This is the "default sequence" we want.
        a<-factor(c(1,2,3),levels=c("3","1","2")). When levels(a), the out put is 3,1,2. This is not "default sequence".
        a<-Transform_fac_num(factor(c(1,2,3),levels=c("3","1","2"))). When levels(a), the output is 1,2,3.
}
\examples{

install.packages("remotes")
library(remotes)
install_gitlab('CarlBrunius/MUVR')
library(MUVR)
factor_1<-factor(c(rep("a",8),rep("b",8),rep("c",8),rep("d",5)))
factor_2<-as.character(c(rep("f",5),rep("e",8),rep("h",5),rep("g",11)))
d<-cbind(factor_1,factor_2,Yotu)
d[29,1]=NA
d[28,2]=NA
d[27,3]=NA
e=Transform_fac_num(d)$dataframe
f<-impute_non_ordinal(e)
f is the imputed data frame of d
}
