% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools_code_OTHER.R
\name{inbra}
\alias{inbra}
\title{Paste objects between brackets}
\usage{
inbra(before = NULL, ..., between = "-", add = NULL,
  rounding = getOption("rounding"))
}
\arguments{
\item{before}{Defaults to NULL. Vector with the values written before the bracket starts. It can also be a data frame (see Details).}

\item{...}{Other vectors to be displayed between brackets.}

\item{between}{Defaults to "-". String to separate values inside the brackets.}

\item{add}{Defaults to NULL. A vector of the strings to add in every term inside the brackets. If it has length 1, it is replicated for all elements.}

\item{rounding}{Defaults to getOption("rounding"). Number of decimals used for rounding. See \code{digits} from \code{\link[base]{round}}.}
}
\value{
Returns a vector of strings with the values before and between the brackets. Has an attribute named \code{data.frame} with the values used.
}
\description{
After the first element, puts everything sorrounded by brackets.
}
\details{
When \code{before} is a data frame, it returns a vector of strings where the first column from the data frame is outside the brackets
and all the others inside separated by the symbol given in \code{between}.
When \code{before} is a vector and \code{...} are also vectors of the same length, a data frame is created and the same that above is performed.
}
\examples{
mins <- apply(iris[1:4], 2, min)
maxs <- apply(iris[1:4], 2, max)
means <- apply(iris[1:4], 2, mean)
inbra(means, mins, maxs, between = " to ")

freqs <- as.numeric(table(iris$Species))
percents <- as.numeric(100*prop.table(table(iris$Species)))
inbra(freqs, percents, rounding = 2, add = "\%")

}
\seealso{
\code{\link{bivarTable}}
}

