% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{k_linear}
\alias{k_linear}
\title{Linear Kernel function}
\usage{
k_linear(x, y)
}
\arguments{
\item{x, }{a numeric}

\item{y, }{a numeric}
}
\value{
a numeric
}
\description{
Given x and y computes the linear kernel and returns it.
}
\examples{
k_linear(c(1,2,3,4), c(3,5,7,9))
}
