% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{K_LLS_R}
\alias{K_LLS_R}
\title{Kernel Regularized Linear Least Squares (K-LLS-R) estimator}
\usage{
K_LLS_R(k, lambda, X = NULL, y = NULL)
}
\arguments{
\item{k, }{a closure}

\item{lambda, }{a numeric}

\item{X, }{a matrix (Optional)}

\item{y, }{a numeric or vector (Optional)}
}
\value{
g, a closure, or if X and Y also supplied a numeric
}
\description{
Given model matrix, X, targets, y, and regularization rate, lambda, and kernel function, k, returns the K-LLS-R estimator.
If only the regularization rate and kernel function is given will return function that computes the predictions using the K-LLS-R estimator for the given regularization rate and kernel function when passed the .
Where our regularization term is "\eqn{\text{lambda} \cdot \mathbf{w}^{T} \mathbf{w}}"
}
\examples{
X <- model_matrix(c(1,2,3,4))
y <- c(1,4,9,16)
k <- k_linear
# The following are equivalent
# --------
f <- K_LLS_R(k, 1)
f(X,y)
# --------
K_LLS_R(k,1,X,y)
}
