% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{k_poly}
\alias{k_poly}
\title{Polynomial Kernel function}
\usage{
k_poly(b, x = NULL, y = NULL)
}
\arguments{
\item{b, }{an integer}

\item{x, }{a numeric}

\item{y, }{a numeric}
}
\value{
a closure, or numeric if x and y also given
}
\description{
Given b will return the polynomial kernel function for degree b, if x and y also given will compute the polynomial kernel (of degree b) and return it.
}
\examples{
The following are the same:
f <- k_poly(4)
f(c(1,2,3,4), c(3,5,7,9))
--------
k_poly(4, c(1,2,3,4), c(3,5,7,9))
}
