% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{k_RBF}
\alias{k_RBF}
\title{RBF Kernel function}
\usage{
k_RBF(sigma, x = NULL, y = NULL)
}
\arguments{
\item{sigma, }{an integer}

\item{x, }{a numeric}

\item{y, }{a numeric}
}
\value{
a closure, or numeric if x and y also given
}
\description{
Given sigma will return the RBF kernel function for bandwidth sigma, if x and y also given will compute the RBF kernel (of bandwidth sigma) and return it.
}
\examples{
# The following are the same:
f <- k_RBF(4)
f(c(1,2,3,4), c(3,5,7,9))
# --------
k_RBF(4, c(1,2,3,4), c(3,5,7,9))
}
