% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattest-functions.R
\name{ks_test}
\alias{ks_test}
\title{Kolmogorov-Smirnov Test for autocorrelated data}
\source{
\code{stats:::ks.test.default}
}
\usage{
ks_test(
  x,
  y,
  a1.x = 0,
  a1.y = 0,
  alternative = c("two.sided", "less", "greater"),
  exact = NULL,
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{y}{a numeric vector of data values.}

\item{a1.x}{autocorrelation parameter at lag one of the data in \code{x}.}

\item{a1.y}{autocorrelation parameter at lag one of the data in \code{y}.}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"two.sided"} (default), \code{"less"}, or
    \code{"greater"}.  You can specify just the initial letter of the
    value, but the argument name must be given in full.
    See \sQuote{Details} for the meanings of the possible values.}

\item{exact}{\code{NULL} or a logical indicating whether an exact
    p-value should be computed.  See \sQuote{Details} for the meaning of
    \code{NULL}.}

\item{simulate.p.value}{a logical indicating whether to compute
    p-values by Monte Carlo simulation.}

\item{B}{an integer specifying the number of replicates used in the
    Monte Carlo test.}
}
\value{
see \code{?ks.test}.
}
\description{
Perform a two-sample Kolmogorov-Smirnov test accounting for autocorrelation
of the input data. This function is a modified version of the base R's
default two-sample \code{\link[stats]{ks.test}}, which treats autocorrelation
after Xu (2013) by resetting the number of observations to \code{n * (1 -a)},
where \code{n} are the original number of observations and \code{a} is the
user-supplied estimate of the data's autocorrelation parameter at lag
one. Note that this code is adapted from the default method of the ks.test
function, thus it can only test whether the data are from the same or from
different distributions and not test the data against a given cumulative
distribution function, nor can it be used with class \code{formula}.
}
\references{
Xu, X.: Methods in Hypothesis Testing, Markov Chain Monte Carlo and
  Neuroimaging Data Analysis, PhD thesis, Harvard University, Cambridge,
  Massachusetts, \url{http://nrs.harvard.edu/urn-3:HUL.InstRepos:11108711},
  2013.
}
\seealso{
\code{\link[stats]{ks.test}}
}
\author{
R Core Team, Thomas Münch
}
