% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LLS}
\alias{LLS}
\title{Linear Least Squares (LLS) estimator}
\usage{
LLS(X, y)
}
\arguments{
\item{X, }{a matrix}

\item{y, }{a numeric or vector}
}
\value{
w, a numeric or vector
}
\description{
Given model matrix, X, and targets, y, returns the LLS estimator
}
\examples{
X <- model_matrix(c(1,2,3,4))
y <- c(1,4,9,16)
LLS(X,y)
}
