% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LLS_R}
\alias{LLS_R}
\title{Regularized Linear Least Squares (LLS-R) estimator}
\usage{
LLS_R(X, y, lambda)
}
\arguments{
\item{X, }{a matrix}

\item{y, }{a numeric or vector}

\item{lambda, }{a numeric}
}
\value{
w, a numeric or vector
}
\description{
Given model matrix, X, targets, y, and regularization rate, lambda, returns the LLS-R estimator.
Where our regularization term is "\eqn{\text{lambda} \cdot \mathbf{w}^{T} \mathbf{w}}"
}
\examples{
X <- model_matrix(c(1,2,3,4))
y <- c(1,4,9,16)
LLS_R(X,y, 1)
}
