% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_worklist.R
\name{merge_worklist}
\alias{merge_worklist}
\title{Please consult King Yan if smth is not clear
One important thing is that you need to set your directory correctly before using the function
You need to double check if your xlsx file name is correct}
\usage{
merge_worklist(Dir = getwd(), File_first, File_second, MS_type = "MS2")
}
\arguments{
\item{File_first}{The first file, POS or NEG, end with xlsx}

\item{File_second}{The first file, POS or NEG, end with xlsx}

\item{MS_type}{you need to specify if the work list is for MS1 or MS2. In most cases MS2}
}
\value{
data
}
\description{
Please consult King Yan if smth is not clear
One important thing is that you need to set your directory correctly before using the function
You need to double check if your xlsx file name is correct
}
