% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{model_matrix}
\alias{model_matrix}
\title{Create a model matrix}
\usage{
model_matrix(D, r, supress = FALSE)
}
\arguments{
\item{D, }{data in the form of a data frame, matrix or vector}

\item{r, }{optional vector of column indices to omit from the model matrix}

\item{supress, }{optional boolean that if TRUE will supress matrix of not full rank warning}
}
\value{
The corresponding model matrix X
}
\description{
Given data, D, and columns to leave out, r, creates the corresponding model matrix.
Note: doesn't notify you or make the model matrix full rank if its not.
}
\examples{
model_matrix(c(1,2,3,4))
model_matrix(cars)
}
