% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Mode (modus)}
\usage{
modus(x)
}
\arguments{
\item{x}{a vector containing the values whose modus is to be
computed.}
}
\value{
A single value containing the most frequent value
 (modus) of \code{x}. For bi- or multimodal distributions
 with equal peaks, the first one is returned (see examples).
}
\description{
Calculate the most frequent value (modus) of a vector.
Missing values (\code{NA}) are ignored.
}
\examples{

x <- round(rnorm(1000), 1)
modus(x)
y <- sample(letters, 100, replace = TRUE)
modus(y)
modus(c(1, 1, 2, 2))
modus(c("b", "b", "a", "a"))
}
\seealso{
\code{\link{median}}, \code{\link{mean}}
}
\author{
Gerhard Nachtmann \email{kpm.nachtmann@gmail.com}
}
\keyword{mode}
\keyword{modus}
