% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvImpWrap.R
\name{mvImpWrap}
\alias{mvImpWrap}
\title{Two-stage random forest imputation}
\usage{
mvImpWrap(MAT, guess = NULL, forceZero = FALSE, method = c("PLS",
  "RF"), rfMeth = c("rf", "Rborist", "ranger"), nComp = 2, nCore,
  tol1 = 0.05, n1 = 15, tol2 = 0.025, n2 = 60)
}
\arguments{
\item{MAT}{Matrix for imputation; samples in rows, variables in columns}

\item{guess}{Initial guess (see mvImp; defaults to mvImp default)}

\item{forceZero}{Boolean for whether to force a lower imputation limit to zero (see mvImp; defaults to mvImp default)}

\item{method}{'PLS' or 'RF}

\item{rfMeth}{Which RF implementation to choose ('rf' (randomForest; default), 'ranger' or 'Rborist')}

\item{nComp}{Number of PLS components (defaults to 2)}

\item{nCore}{Number of slave processes (defaults to detectCores()-1)}

\item{tol1}{Tolerance in 1st iteration (defaults to 0.05)}

\item{n1}{MaxIter for 1st iteration (defaults to 15)}

\item{tol2}{Tolerance in 2nd iteration (defaults to 0.05)}

\item{n2}{MaxIter for 2nd iteration (defaults to 15)}
}
\value{
A data matrix with imputed values
}
\description{
This wrapper performs parallel processing to make a 2-stage rfImp
1st stage has "sloppy" settings to provide a reasonable 1st guess
for the 2nd iteration which has "nicer" settings.
}
\examples{
MAT=matrix(runif(100000),nrow=40) # Generate synthetic (random) data
MAT[sample(1:100000,size = 1000)] <- NA # Punch 1000 random holes in the data
MAT_Imp <- rfImpWrap(MAT = MAT) # Imputation using default values
}
