% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_check.R
\name{overlapping_check}
\alias{overlapping_check}
\title{This function is to check for overlapping of values or dataframe columns in different objects.
You could see this as a venn diagram but in nicer text version. It does not have the limitation of venn diagram that only at most 5-6 object can be shown together}
\usage{
overlapping_check(X)
}
\arguments{
\item{X}{This should be a list of vectors or dataframes (There must be names of each object in the list). 
Each object in the list should be in the same format (all in vector or all in data.frame).  
If it is a list of dataframe, then the rows should be observations and columns variables, 
and only completely identical columns (both column names and values) will be considered as overlapped
If it is a list of dataframe, the columns with the same names accross object should have exactly identical vaule,
otherwise an eroor is given (Thus this function also serve the purpose of check if data are identical or not)}
}
\value{
all the "and" and "or"options
}
\description{
This function is to check for overlapping of values or dataframe columns in different objects.
You could see this as a venn diagram but in nicer text version. It does not have the limitation of venn diagram that only at most 5-6 object can be shown together
}
