% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCor.R
\name{pCor}
\alias{pCor}
\title{This function aims to construct one more flexible pcor()}
\usage{
pCor(X, Y, C, cor_method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{X}{A data frame of X variables (All variables needs to be numeric)}

\item{Y}{A vector/dataframe of one Y variable (needs to be numeric)}

\item{C}{A data frame of covariates (All variables needs to be numeric)}

\item{cor_method}{from cor()}
}
\value{
A vector with adjusted partial correlations
}
\description{
This function aims to construct one more flexible pcor()
}
