% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.R
\name{plotDens}
\alias{plotDens}
\title{Make XY scatterplot with X and Y histograms and densities on the margins}
\usage{
plotDens(x, y, xlab, ylab, export = FALSE, breaks = 30, ...)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector ofsame length as x}

\item{xlab}{Optional x label (otherwise defaults to the name of x)}

\item{ylab}{Optional y label (otherwise defaults to the name of y)}

\item{export}{Option to export to png or pdf}

\item{breaks}{number of breaks in the margin histograms}

\item{...}{}
}
\value{
A 2D plot
}
\description{
Make XY scatterplot with X and Y histograms and densities on the margins
}
\examples{
randx <- runif(100)
randy <- runif(100)
plotDens(randx, randy)
}
