% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.R
\name{PopSD}
\alias{PopSD}
\title{Population standard deviation}
\usage{
PopSD(x)
}
\arguments{
\item{x}{numeric; vector of values.}
}
\value{
The population standard deviation value of \code{x}.
}
\description{
Calculate the population standard deviation.
}
\examples{

# analytical solution to SD of uniform distribution:
sqrt(1/12 * 10^2)

sd(1:10)

PopSD(1:10)
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
