% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools_code_OTHER.R
\name{repath}
\alias{repath}
\title{Windows path to R string}
\usage{
repath(writecb = TRUE)
}
\arguments{
\item{writecb}{Defaults to TRUE. If TRUE, the changed path to be read in R will be writted in the clipboard so that it can
be pasted in R. Done via}
}
\value{
Returns a string with the new path adapted for R or the working directory if it did not exist.
}
\description{
Reads the clipboard (with the copied windows path) and returns (and writes) in it a string to be pasted in an R environment.
}
\details{
repath() has only two parameters. One is \code{write}, and the other is the windows path already copied in the clipboard. If 
the copied path does not exist, the working directory is returned with a warning.
}
\examples{
\dontrun{
# copy an R path from Windows and do 
repath()
# now paste where you need the adapted path. Voila!
}
}

