% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWeights.R
\name{RWeights}
\alias{RWeights}
\title{Draw random number between 1 and length(weights) with the distribution}
\usage{
RWeights(n = 1000, weights)
}
\arguments{
\item{n}{Number of random numbers to be drawn}

\item{weights}{vector of numbers; defines the distribution}
}
\value{
vector of n random numbers between 1 and length(weights)
}
\description{
Draw random number between 1 and length(weights) with the distribution
}
\examples{
weights <- c(rep(1,10), 1:10, 10:1)
hist(RWeights(100000, weights), breaks = 501)
}
\author{
Thomas Laepple
}
