% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file_names_copy_to_excel.R
\name{scan_file_names_copy_to_excel}
\alias{scan_file_names_copy_to_excel}
\title{Scan file names, copy to excel}
\usage{
scan_file_names_copy_to_excel(Dir, full_names = F, keyphrase, rerank = T)
}
\arguments{
\item{Dir}{directory where the folder is. This folder contains folder where contain all files with names you want to scan}

\item{full_names}{logical, If you want the file names output with the directory names or not}

\item{keyphrase}{a vector of key phrase in the file names that separate the columns. Case sensitive}

\item{rerank}{logical. Rerank base on the number at the}
}
\value{
a excel file and a dataframe
}
\description{
Scan file names, copy to excel
}
