% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-seasonal-temperatures.R
\name{SeasonalCycle}
\alias{SeasonalCycle}
\alias{dSeas}
\alias{rSeas}
\title{Temperature distribution of seasonal cycle}
\usage{
dSeas(x = NULL, mean.t, amp.t, sd.t, return = c("density", "FUN"), res = 0.01)

rSeas(n, mean.t, amp.t, sd.t)
}
\arguments{
\item{x}{numeric; vector of temperatures.}

\item{mean.t}{numeric; mean annual temperature.}

\item{amp.t}{numeric; peak-to-peak amplitude of the seasonal cycle in
temperature.}

\item{sd.t}{numierc; inter-annual standard deviation of the temperature at a
given point in the seasonal cycle.}

\item{return}{return either the densities for the temperatures \code{x}, or a
function to do so; default: c("density", "FUN").}

\item{res}{numeric; resolution of the temperature vectors to be convolved;
default is 0.01.}

\item{n}{integer; number of samples.}
}
\description{
Probability density function and random number generation for the temperature
distribution in the seasonal cycle.
}
\details{
The seasonal temperature cycle is approximated as a sine wave with gaussian
noise. The probability density function is obtained by numerically
convolving a scaled sine wave with a normal PDF with mean equal to 0. Random
deviates are generated by adding gaussian noise to points sampled uniformly
from the scaled sine wave on the interval 0 to 2*pi.
}
\examples{
\dontrun{
if(interactive()){
amp.t <- 10
mean.t <- 25
sd.t <- 0.75

x <- seq(mean.t - amp.t, mean.t + amp.t, length.out = 1000)
Z <- rSeas(n = 100000, mean.t, amp.t, sd.t)

hist(Z, freq=F,breaks=100)
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.1), col = "Blue")
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.01), col = "Green")
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.001), col = "Red")
 }
}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
