% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-functions.R
\name{ShiftCorrelation}
\alias{ShiftCorrelation}
\title{Lag correlation analysis}
\usage{
ShiftCorrelation(
  y1,
  y2,
  shifts,
  unit = NULL,
  plot = FALSE,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{y1}{numeric vector of first data series.}

\item{y2}{numeric vector of second data series which is shifted relative to
\code{y1}; needs to have the same length as \code{y1}.}

\item{shifts}{integer vector of shifts (can be positive, negative, or
both). For each value, \code{y2} is shifted the corresponding number of
elements earlier or later, and the correlation with \code{y1} is computed.}

\item{unit}{optional value for the unit which matches the shifts to the
physical scale on which the data is tabulated.}

\item{plot}{logical; plot an optional graph of correlation versus tried
shifts.}

\item{xlab}{optional x axis label for the plot.}

\item{ylab}{optional y axis label for the plot.}
}
\value{
Data frame with one row with the optimal shift and the corresponding
  correlation value; if a \code{unit} is provided, it includes an additional
  column with the optimal shift in the real units.
}
\description{
Find the optimal shift (positive or negative) between two data vectors from
maximising the correlation within their common overlap.
}
\examples{
y1 <- rnorm(100)
y2 <- rnorm(100)

ShiftCorrelation(y1, y2, shifts = (-10 : 10))

# display optimal shift in the real units of the data:
ShiftCorrelation(y1, y2, shifts = (-10 : 10), unit = 3)
}
\author{
Thomas Münch
}
