% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(z)
}
\arguments{
\item{z, }{a numeric}
}
\value{
a numeric
}
\description{
Outputs the value of the sigmoid function for a given value, z.
}
\examples{
z <- runif(1)
sigmoid(z)
}
