% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_randomly.R
\name{split_randomly}
\alias{split_randomly}
\title{split randomly}
\usage{
split_randomly(X, Y, k)
}
\arguments{
\item{X}{a dataframe}

\item{Y}{a vector}

\item{k}{k fold}
}
\value{
split
}
\description{
split randomly
}
