% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_traintest.R
\name{split_traintest}
\alias{split_traintest}
\title{split traintest}
\usage{
split_traintest(X, Y, ratio)
}
\arguments{
\item{X}{a dataframe}

\item{Y}{a vector}

\item{ratio}{ratio}
}
\value{
data
}
\description{
split traintest
}
