% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitsamplefeatures.R
\name{splitsamplefeatures}
\alias{splitsamplefeatures}
\title{split sample features}
\usage{
splitsamplefeatures(
  sampleNames,
  sampleFeatures,
  rententiontime_threshold = 20,
  samplesplitlistmark = "£"
)
}
\arguments{
\item{sampleNames}{A vector with sample names}

\item{sampleFeatures}{A series of list that we want to investigate}

\item{rententiontime_threshold}{the difference of time point that we want to specify}
}
\value{
A split list
}
\description{
split sample features
}
