% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools.R
\docType{package}
\name{statTools}
\alias{statTools}
\alias{statTools-package}
\title{statTools}
\description{
Useful functions for performing simple statistical analyses, particularly some plots and predefined tables with results.
}

