% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_split.R
\name{stratified_split}
\alias{stratified_split}
\title{Stratified split}
\usage{
stratified_split(X, Y, k)
}
\arguments{
\item{X}{a dataframe}

\item{Y}{a vector}

\item{k}{kfold}
}
\value{
split
}
\description{
Stratified split
}
