% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-functions.R
\name{SummariseEmpiricalPDF}
\alias{SummariseEmpiricalPDF}
\title{Summarise empirical distribution}
\usage{
SummariseEmpiricalPDF(x, p)
}
\arguments{
\item{x}{numeric; vector of values of the empirical PDF.}

\item{p}{numeric; vector of probabilities of the empirical PDF.}
}
\value{
Returns a named vector with the mean, median, mode, and standard
deviation of the empirical PDF.
}
\description{
Summarise an empirical probability distribution function.
}
\details{
Calculation of the mode is naive. For a multimodal distribution only the
highest is returned, in the case of 2 or more modes with exactly the same
probability, the first is returned.
}
\examples{
df <- data.frame(x = 1:10)
df$p <- dnorm(df$x, 5, 2)
SummariseEmpiricalPDF(df$x, df$p)
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
