% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stuff.R
\name{summary_stuff}
\alias{summary_stuff}
\title{Summary stuff}
\usage{
summary_stuff(V, type = c("mean", "sd", "variance", "table"), ID)
}
\arguments{
\item{V}{variable for all observations}

\item{type}{mean, standard deviation,variance, standard error or table}

\item{ID}{unique ID number is smaller than number of observations}
}
\value{
summary statistics
}
\description{
Summary stuff
}
