% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swithchNames.R
\name{switchNames}
\alias{switchNames}
\title{Switch values in vector}
\usage{
switchNames(vector, old, new)
}
\arguments{
\item{vector}{Vector where you want to switch from one set of values to another}

\item{old}{Vector with "levels" in the old format}

\item{new}{Vector with "levels" in the new format}
}
\value{
Vector with new values
}
\description{
Switch values in vector
}
