% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattest-functions.R
\name{t_test}
\alias{t_test}
\title{Student's t-Test for autocorrelated data}
\source{
\code{stats:::t.test.default}
}
\usage{
t_test(
  x,
  y = NULL,
  a1.x = 0,
  a1.y = 0,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{a1.x}{autocorrelation parameter at lag one of the data in \code{x}.}

\item{a1.y}{autocorrelation parameter at lag one of the data in \code{y}.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}.  You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in
means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var.equal}{a logical variable indicating whether to treat the two
variances as being equal. If ‘TRUE’ then the pooled variance is used to
estimate the variance otherwise the Welch (or Satterthwaite) approximation
to the degrees of freedom is used.}

\item{conf.level}{confidence level of the interval.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
see \code{?t.test}.
}
\description{
Perform a one-sample or a two-sample t-test accounting for autocorrelation of
the input data. This function is a modified version of the base R's default
\code{\link[stats]{t.test}}, which treats autocorrelation of the input data
by resetting the number of observations to the effective degrees of freedom
according to user-supplied estimates of the data's autocorrelation parameters
at lag one. Note that this code is adapted from the default method of the
t.test function, thus it cannot be used with class \code{formula}.
}
\seealso{
\code{\link[stats]{t.test}}, \code{\link{getEffectiveDOF}}
}
\author{
R Core Team, Thomas Münch
}
