% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_fac_num.R
\name{Transform_fac_num}
\alias{Transform_fac_num}
\title{Transform all variable class as either numeric or factor.
Make the sequence of the levels for each factor variable become the "default sequence".}
\usage{
Transform_fac_num(X)
}
\arguments{
\item{X}{A data frame with variables with class "numeric", "factor", "logical","character","ordered factor"}
}
\value{
A list with 2 items: A dataframe with all variable class as either numeric or factor; Names of ordinal factors that is originally classified as "ordered factor".
}
\description{
Example: a<-factor(c(1,2,3)). When levels(a), the out put is 1,2,3. This is the "default sequence" we want.
        a<-factor(c(1,2,3),levels=c("3","1","2")). When levels(a), the out put is 3,1,2. This is not "default sequence".
        a<-Transform_fac_num(factor(c(1,2,3),levels=c("3","1","2"))). When levels(a), the output is 1,2,3.
}
\examples{
#install.packages("remotes")
#library(remotes)
#install_gitlab('CarlBrunius/MUVR')
#factor_1<-factor(c(rep("2",8),rep("3",8),rep("4",8),rep("1",5)),levels=c("3","1","2","4"))
#factor_2<-as.character(c(rep("f",5),rep("e",8),rep("h",5),rep("g",11)))
#d<-cbind(Xotu[,1:7],factor_1,factor_2,Yotu)
#e=Transform_fac_num(d)$dataframe


}
