% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\title{Plot Venn diagram}
\usage{
venn(x, ...)
}
\arguments{
\item{x}{Same as for the `venn.diagram()` function, i.e. a list of names}

\item{...}{Standard `venn.diagram()` arguments}
}
\description{
Convenience function to draw a Venn diagram directly in the plot window without exporting to a file
}
\examples{
StatTools::venn(x = list(set1 = sample(letters)[1:10], set1 = sample(letters)[1:15], set1 = sample(letters)[1:20]))
}
