% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnDec.R
\name{AnDec}
\alias{AnDec}
\title{ANOVA decomposition}
\usage{
AnDec(inData, levels)
}
\arguments{
\item{inData}{A matrix/data.frame with inData (samples as rows, variables as columns)}

\item{levels}{A matrix/data.frame with levels (one factor per column, multiple levels per factor)}
}
\value{
An object containing ANOVA-decompositioned data
}
\description{
ANOVA decomposition of `inData` according to `levels`
}

