% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brs.R
\name{brs}
\alias{brs}
\title{Summary table for bivariate regression models}
\usage{
brs(..., EXP  =  TRUE, EST, Out)
}
\arguments{
\item{...}{List of the bivariate regression models separated by comma}

\item{EXP}{Logical, whether to exponentiate the estimate or not}

\item{EST}{What to call the estimate column (a quoted string e.g. OR, HR)}

\item{Out}{The label for the response variable (a quoted string)}
}
\description{
This function creates a summary table from multiple
 bivariate regression models. If you would like to examine the bivariate
 associations between multiple potential confounders and an outcome variable,
 you may want to run multiple bivariate models and summarize 
 the results in one table. This function helps you with that.
 brs: stands for bivariate regression summary.
}
\examples{
# Example
set.seed(1234)
model1 <- glm(vs ~ gear, data = mtcars, family = binomial)
model2 <- glm(vs ~ hp, data = mtcars, family = binomial)
model3 <- glm(vs ~ drat, data = mtcars, family = binomial)
brs(model1,model2, model3,EXP = TRUE, Out = "Outcome:VS", EST = "OR")
}
