% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-seasonal-temperatures.R
\name{dSin}
\alias{dSin}
\title{Density of a scaled sin wave}
\usage{
dSin(x, mean.t, amp.t)
}
\arguments{
\item{mean.t}{mean of sine wave}

\item{amp.t}{amplitude of sine wave}
}
\description{
Density of a scaled sin wave
}
