% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_all.R
\name{impute_all}
\alias{impute_all}
\title{imputation for 3 types of variables
1.numeric variables
2.factor variables without order: non_ordinal factor variables
3.factor variable with order: ordinal factor variables}
\usage{
impute_all(ordinal, non_ordinal)
}
\arguments{
\item{ordinal}{A dataframe with ordinal variables and numeric variables}

\item{non-ordinal}{A dataframe with non-ordinal variables}
}
\value{
A dataframe with imputed values
}
\description{
Preparation: To use this package, 4 steps need to be done beforehand:
}
\details{
Step 1: Made all variables that will be used a dataframe. For example, a data frame with name X

Step 2: Put ordinal variables and numeric variables in one dataframe. Put  non-ordinal variables in another data frame.
This is done manual because computer cannot decide which variables we vant to give order to.
!!! Note: the class of variables in these data frames should be only "factor" or "numeric". (Use the Transform_fac_num in the same package.)

Step 3: Code the ordinal factor variables properly!!!. For example, code it with alphabet or number that could show sequence.

Step 4: Make sure that the sequence of the levels for each factor variable become the "default sequence". (Use the Transform_fac_num in the same package.)
Example: a<-factor(c(1,2,3)). When levels(a), the out put is 1,2,3. This is the "default sequence" we want.
        a<-factor(c(1,2,3),levels=c("3","1","2")). When levels(a), the out put is 3,1,2. This is not "default sequence".
        a<-Transform_fac_num(factor(c(1,2,3),levels=c("3","1","2"))). When levels(a), the output is 1,2,3.
}
