% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_non_ordinal.R
\name{impute_non_ordinal}
\alias{impute_non_ordinal}
\title{imputation for non-ordinal factor variables}
\usage{
impute_non_ordinal(non_ordinal)
}
\arguments{
\item{non-ordinal}{A dataframe with non-ordinal variables}
}
\value{
A dataframe with imputed values

install.packages("remotes")
library(remotes)
install_gitlab('CarlBrunius/MUVR')
library(MUVR)
factor_1<-factor(c(rep("a",8),rep("b",8),rep("c",8),rep("d",5)))
factor_2<-as.character(c(rep("f",5),rep("e",8),rep("h",5),rep("g",11)))
d<-cbind(factor_1,factor_2,Yotu)
d[29,1]=NA
d[28,2]=NA
d[27,3]=NA
e=Transform_fac_num(d)$dataframe
f<-impute_non_ordinal(e)
# f is the imputed data frame of d
}
\description{
Preparation: To use this package, 1 steps need to be done beforehand:
}
\details{
Step 1: Made all variables that will be used a dataframe. For example, a data frame with name X

!!! Note: the class of variables in these data frames should be only "factor". (Use the Transform_fac_num in the same package.)

Step 2:  Make sure that the sequence of the levels for each factor variable become the "default sequence". (Use the Transform_fac_num in the same package.)
Example: a<-factor(c(1,2,3)). When levels(a), the out put is 1,2,3. This is the "default sequence" we want.
        a<-factor(c(1,2,3),levels=c("3","1","2")). When levels(a), the out put is 3,1,2. This is not "default sequence".
        a<-Transform_fac_num(factor(c(1,2,3),levels=c("3","1","2"))). When levels(a), the output is 1,2,3.
}
