% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTools_code_OTHER.R
\name{inbra}
\alias{inbra}
\title{Paste objects between brackets}
\usage{
inbra(before, ..., between = "-", add = "",
  rounding = getOption("rounding"))
}
\arguments{
\item{before}{Vector with the values written before the bracket starts. It can also be a data frame or matrix (see Details).}

\item{...}{Other vectors to be displayed between brackets.}

\item{between}{Defaults to "-". Character vector with the strings to separate values inside the brackets. If it has length 1, it is replicated for all elements.}

\item{add}{Defaults to \code{""}. A character vector with strings to add after every value inside the brackets. If it has length 1, it is replicated for all elements.}

\item{rounding}{Defaults to getOption("rounding"). Number of decimals used for rounding. See \code{digits} from \code{\link[base]{round}}.}
}
\value{
Returns a vector of strings with the values before and between the brackets. Has an attribute named \code{data.frame} with the values used.
}
\description{
The first element is placed outside brackets and the rest inside. In de form of \code{before (inside1, inside2, ...)}
}
\details{
When \code{before} is a data frame, it returns a vector of strings where the first column from the data frame is outside the brackets
and all the others inside separated by the symbol given in \code{between}.
When \code{before} is a vector and \code{...} are also vectors of the same length, a data frame is created and the same that above is performed.
}
\examples{
mins <- apply(iris[1:4], 2, min)
maxs <- apply(iris[1:4], 2, max)
means <- apply(iris[1:4], 2, mean)
inbra(means, mins, maxs, between = " to ")

freqs <- as.numeric(table(iris$Species))
percents <- as.numeric(100*prop.table(table(iris$Species)))
inbra(freqs, percents, rounding = 2, add = "\%")

# with a data frame:
values <- inbra(iris[c(5, 1:4)], between = c("; ", " - ", " - "), add = "cm")
values

# data arguments are still stored in
attributes(values)$data

}
\seealso{
\code{\link{bivarTable}}
}

