% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LLS_R}
\alias{LLS_R}
\title{Regularized Linear Least Squares (LLS-R) estimator}
\usage{
LLS_R(lambda, X = NULL, y = NULL)
}
\arguments{
\item{lambda, }{a numeric}

\item{X, }{a matrix (Optional)}

\item{y, }{a numeric or vector (Optional)}
}
\value{
w, a numeric or vector
}
\description{
Given model matrix, X, targets, y, and regularization rate, lambda, returns the LLS-R estimator.
If only the regularization rate is given will return function that computes LLS-R for the given regularization rate.
Where our regularization term is "\eqn{\text{lambda} \cdot \mathbf{w}^{T} \mathbf{w}}"
}
\examples{
X <- model_matrix(c(1,2,3,4))
y <- c(1,4,9,16)
# The following are equivalent
--------
f <- LLS_R(1)
f(X,y)
--------
LLS_R(1,X,y)
}
