% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-functions.R
\name{PopSD}
\alias{PopSD}
\title{Population standard deviation}
\usage{
PopSD(x)
}
\arguments{
\item{x}{}
}
\value{
numeric, population standard deviation
}
\description{
Population standard deviation
}
\examples{

# analytical solution to SD of uniform
sqrt(1/12 * 10^2)
 
sd(1:10)

PopSD(1:10)
}
