% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfImp.R
\name{rfImp}
\alias{rfImp}
\title{Perform random forest imputation of missing values}
\usage{
rfImp(MAT, maxIter = 15, tolerance = 0.01, ntree = 100, mtry = 5,
  guess, verbose = F, parallel = F)
}
\arguments{
\item{MAT}{Indata matrix}

\item{maxIter}{Maximum number of iterations}

\item{tolerance}{Tolerance criterion (between two iterations) to accept imputed value}

\item{ntree}{Number of trees per forest}

\item{mtry}{Number of variables to choose from at each node}

\item{guess}{1st guess ("median", i.e. median per feature; "minVar", minimum per variable (default); "minTotal", minimum of all variables; A matrix (imputed/full data matrix))}

\item{verbose}{Boolean for verbose output}

\item{parallel}{Whether to perform imputation in parallel (register backend if TRUE!)}
}
\value{
Object (list)
}
\description{
Perform random forest imputation of missing values
}
