% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWeights.R
\name{RWeights}
\alias{RWeights}
\title{Draw random number from distribution given the weights}
\usage{
RWeights(n = 1000, weights)
}
\arguments{
\item{n}{integer; number of random integers to be drawn.}

\item{weights}{numeric vector of weights which define the distribution.}
}
\value{
Vector of \code{n} random numbers between 1 and the length of
  \code{weights}.
}
\description{
Draw a given number of random integer values with replacement from a
distribution given by its weights, where the values can range between 1 and
the length of the weights vector.
}
\examples{
weights <- c(rep(1,10), 1:10, 10:1)
hist(RWeights(100000, weights), breaks = 501)
}
\author{
Thomas Laepple
}
