% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.R
\name{SDofSD}
\alias{SDofSD}
\title{Standard deviation (error) of standard deviation}
\usage{
SDofSD(s, n)
}
\arguments{
\item{s}{numeric; estimated standard deviation.}

\item{n}{integer; number of samples from which \code{s} was estimated.}
}
\value{
Numeric value giving the standard error of the estimated standard
  deviation \code{s}.
}
\description{
Calculate the standard deviation (i.e. standard error) of a standard
deviation, given n, the number of samples from which the standard deviation
was estimated.
}
\details{
If n <= 300, an exact method is used, otherwise Stirling's approximation is
used to avoid numerical errors calculating the gamma function for large
values of n. Formulae taken from https://stats.stackexchange.com/q/28567.
}
\examples{
\dontrun{
if(interactive()){
 SDofSD(1, 12)
 library(ggplot2)
df <- expand.grid(SD = 1:10, n = seq(2, 10, length.out = 100))
df$SDofSD = SDofSD(df$SD, df$n)
ggplot(df, aes(x = n, y = SDofSD, group = SD, colour = SD)) +
  geom_line()
 }
}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
