% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-seasonal-temperatures.R
\name{SeasonalCycle}
\alias{SeasonalCycle}
\alias{dSeas}
\alias{rSeas}
\title{Distribution of temperatures in seasonal cycle}
\usage{
dSeas(x = NULL, mean.t, amp.t, sd.t, return = c("density", "FUN"), res = 0.01)

rSeas(n, mean.t, amp.t, sd.t)
}
\arguments{
\item{x}{vector of temperatures}

\item{mean.t}{mean annual temperature}

\item{amp.t}{amplitude of the seasonal cycle}

\item{sd.t}{inter-annual standard deviation of temperature at a given point
in the seasonal cycle}

\item{return}{return densities for temperatures x, or a function to do so,
Default: c("density", "FUN")}

\item{res}{resolution of the temperature vectors to be convolved, Default:
0.01}

\item{n}{number of samples}
}
\description{
Probability density function and random generation for seasonal
  temperatures
}
\details{
The seasonal temperature cycle is approximated as a sine wave with
  gaussian noise. The probability density function is obtained by numerically
  convoluting a scaled sine wave and normal PDF with mean = 0. Random
  deviates are generated by adding gaussian noise to points sampled uniformly
  from the scaled sine wave on the interval 0 to 2*pi
}
\examples{
\dontrun{
if(interactive()){
amp.t <- 10
mean.t <- 25
sd.t <- 0.75

x <- seq(mean.t-amp.t, mean.t+amp.t, length.out = 1000)
Z <- rSeas(n = 100000, mean.t, amp.t, sd.t)

hist(Z, freq=F,breaks=100)
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.1), col = "Blue")
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.01), col = "Green")
lines(x, dSeas(x, mean.t, amp.t, sd.t, res = 0.001), col = "Red")
 }
}
}
\author{
Andrew Dolman
}
